+++ csense-kotlin-test-common/src/main/kotlin/csense/kotlin/test/Async.kt
@@ -1,6 +1,6 @@
 package csense.kotlin.test
 
-import kotlinx.coroutines.experimental.*
+import kotlinx.coroutines.*
 
 fun testAsync(method: suspend () -> Unit) {
     GlobalScope.launch(Dispatchers.Unconfined) {